requirements.txt
einops==0.8.1
encodec==0.1.1
huggingface_hub==0.23.0
jsonargparse==4.28.0
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning_utilities==0.11.9
mamba_ssm==1.2.0.post1
matplotlib==3.8.0
numpy==2.2.3
onnxruntime_gpu==1.18.0
pandas==2.2.3
pesq==0.0.4
pytorch_lightning==2.0.3
#PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
soundfile==0.12.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchinfo==1.8.0
torchmetrics==1.1.2
tqdm==4.66.4
transformers==4.40.1
